antigravity/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy
opencv-python
pillow
tifffile
google-generativeai
pandas
